completed course: adding cover image to cover image

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,23 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     * for a user to access the post manipulation pages
+     * they have to be authenticated otherwise, BLOCKEDT.
+     * except some views- index and show
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -46,14 +59,38 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' =>'required'
+            'body' =>'required',
+            'cover_image' => 'image|nullable|max:1999'
+            //should be an image, can be nullable and max is 1999 bytes - max for
+            //apache servers is 2mb
         ]);
+
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //get file name with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //just get filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //just get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            //cover_images folder stored in storage/app/public - inaccessible through browser
+            //set symlink to public folder through: php artisan storage:link
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+
+        }else{ //if no image is uploaded
+            $filenameToStore = 'noimage.jpg';
+
+        }
 
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image =  $filenameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -80,6 +117,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //prevents unauthorised users from editing and deletingby redirecting
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -94,13 +137,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' =>'required'
+            'body' =>'required',
         ]);
 
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //get file name with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //just get filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //just get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            //cover_images folder stored in storage/app/public - inaccessible through browser
+            //set symlink to public folder through: php artisan storage:link
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+
+        }
         //Update Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $filenameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -116,6 +178,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            //delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
